Extract apigateway helper functions
